refactor: move rapepr to cmponents, added the review and tested toggle
